+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/DefineSpriteTag.java
@@ -31,6 +31,7 @@ import com.jpexs.decompiler.flash.tags.base.PlaceObjectTypeTag;
 import com.jpexs.decompiler.flash.tags.base.RemoveTag;
 import com.jpexs.decompiler.flash.tags.base.RenderContext;
 import com.jpexs.decompiler.flash.tags.base.SoundStreamHeadTypeTag;
+import com.jpexs.decompiler.flash.tags.gfx.DefineExternalStreamSound;
 import com.jpexs.decompiler.flash.timeline.Timeline;
 import com.jpexs.decompiler.flash.timeline.Timelined;
 import com.jpexs.decompiler.flash.types.BasicType;
@@ -369,7 +370,7 @@ public class DefineSpriteTag extends DrawableTag implements Timelined {
     @Override
     public void getNeededCharacters(Set<Integer> needed) {
         for (Tag t : getTags()) {
-            if ((t instanceof CharacterIdTag) && !(t instanceof SoundStreamHeadTypeTag)) {
+            if ((t instanceof CharacterIdTag) && !(t instanceof SoundStreamHeadTypeTag) && !(t instanceof DefineExternalStreamSound)) {
                 needed.add(((CharacterIdTag) t).getCharacterId());
             }
         }